# Add Benchmark and minor performance refactor