--- a/ClosedXML.Report.sln
+++ b/ClosedXML.Report.sln
@@ -11,6 +11,8 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "tests", "tests", "{8438BAA8
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ClosedXML.Report.Tests", "tests\ClosedXML.Report.Tests\ClosedXML.Report.Tests.csproj", "{D529A371-3AEA-4D02-9A0D-308A6276D411}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ClosedXML.Report.Benchmarks", "ClosedXML.Report.Benchmarks\ClosedXML.Report.Benchmarks.csproj", "{E61B9B2F-96AF-4B19-9327-C7B724BA0B1D}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -25,6 +27,10 @@ Global
 		{D529A371-3AEA-4D02-9A0D-308A6276D411}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{D529A371-3AEA-4D02-9A0D-308A6276D411}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{D529A371-3AEA-4D02-9A0D-308A6276D411}.Release|Any CPU.Build.0 = Release|Any CPU
+		{E61B9B2F-96AF-4B19-9327-C7B724BA0B1D}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{E61B9B2F-96AF-4B19-9327-C7B724BA0B1D}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{E61B9B2F-96AF-4B19-9327-C7B724BA0B1D}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{E61B9B2F-96AF-4B19-9327-C7B724BA0B1D}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -32,6 +38,7 @@ Global
 	GlobalSection(NestedProjects) = preSolution
 		{DEF4A219-D1CF-42A2-A5AC-FE7F2F005AB0} = {885EF47A-C124-45F0-B979-303B30FB586B}
 		{D529A371-3AEA-4D02-9A0D-308A6276D411} = {8438BAA8-E968-4C14-9500-E3A52434E32A}
+		{E61B9B2F-96AF-4B19-9327-C7B724BA0B1D} = {8438BAA8-E968-4C14-9500-E3A52434E32A}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {42DA0EC5-D733-4746-9613-413A64EEDC4B}