--- a/ClosedXML.Report/Excel/CellPosition.cs
+++ b/ClosedXML.Report/Excel/CellPosition.cs
@@ -1,0 +1,14 @@
+namespace ClosedXML.Report.Excel;
+
+struct CellPosition
+{
+    public CellPosition(int row, int column)
+    {
+        Row = row;
+        Column = column;
+    }
+
+    public int Row { get; set; }
+
+    public int Column { get; set; }
+}