… ClosedXML.Report.Benchmarks/Resources/Benchmark.xlsx
… tests/Gauges/tLists7_horizontal_images.xlsx