### Benchmarking Project Setup:
* Added a new benchmarking project `ClosedXML.Report.Benchmarks`  to measure report generation performance using the BenchmarkDotNet library. (`ClosedXML.Report.Benchmarks/Benchmarks/ReportBenchmarks.cs`, [ClosedXML.Report.Benchmarks/Benchmarks/ReportBenchmarks.csR1-R46](diffhunk://#diff-33c77b0f536df0a6b2415223ccdfb32a4e6e6bbfdb35699e2f70d5d9caa426e1R1-R46))

### Refactoring for Maintainability:
* Refactored `TempSheetBuffer` to replace individual row and column variables with a `CellPosition` struct, simplifying cell position management and improving code readability. (`ClosedXML.Report/Excel/TempSheetBuffer.cs`, [ClosedXML.Report/Excel/TempSheetBuffer.csL13-R135](diffhunk://#diff-4474c19cd6c40fb06da81951e910610c545250a44fd7e347d831d8a476a4a477L13-R135))
* Updated `SubtotalSummaryFunc` to use modern C# features such as target-typed `new()` and pattern matching, and replaced verbose property definitions with expression-bodied members. (`ClosedXML.Report/Excel/SubtotalSummaryFunc.cs`, [[1]](diffhunk://#diff-de8c42538de54280696159e9b6388633e56221d77b60dbed7050760b2f19285fL4-R10) [[2]](diffhunk://#diff-de8c42538de54280696159e9b6388633e56221d77b60dbed7050760b2f19285fL30-R29) [[3]](diffhunk://#diff-de8c42538de54280696159e9b6388633e56221d77b60dbed7050760b2f19285fL174-R190)

### Motivations

We have code that generates spreadsheets, some are taking an increasing amount of time and memory to create.  A lot of the performance is in the underlying ClosedXml, not so much reporting lib, but these changes add a slight drop in raw and even slighter performance boost in general. 

#### Benchmark this PR

| Method           |    Mean |    Error |   StdDev |   Gen0 |  Gen1 | Gen2 | Allocated |
| ---------------- | ------: | -------: | -------: | -----: | ----: | ---: | --------: |
| ReportGeneration | 2.867 s | 0.0552 s | 0.0737 s | 131000 | 45000 | 5000 |   2.04 GB |
| ReportGeneration | 2.753 s | 0.0534 s | 0.0500 s | 131000 | 45000 | 5000 |   2.04 GB |

### Benchmark Existing code

| Method           |    Mean |    Error |   StdDev |   Gen0 |  Gen1 | Gen2 | Allocated |
| ---------------- | ------: | -------: | -------: | -----: | ----: | ---: | --------: |
| ReportGeneration | 2.940 s | 0.0545 s | 0.0968 s | 134000 | 45000 | 5000 |   2.08 GB |
| ReportGeneration | 2.902 s | 0.0570 s | 0.0679 s | 134000 | 44000 | 5000 |   2.08 GB |

Flamegraph shows a lot of time is spent in the underlying ClosedXML lib, one hot path for us is in conditional formatting application.  I'll see if I can review that in the ClosedXML project.

![image](https://github.com/user-attachments/assets/82426148-7d51-4b26-a3f8-e6aa87ffde92)

## Test changes

I amended one expected test, it was failing I Think the change is justified.
Here is the existing test expectation

![image](https://github.com/user-attachments/assets/a619e55b-b9e4-4c72-898f-2502420ed9b5)

Here is the amended: note the formatting on the `Item 6 cell`.  All the other cells follow this pattern, so I think it made sense that this cell also should be rendered the same. 
![image](https://github.com/user-attachments/assets/e0ae8df8-ed78-4fe7-8f20-95370bbfaf90)

I actually think all the cells should have the border formatting, but that is for another day.